fix(webui): extract image resize function into function file and make size a parameter
